requirements.txt
PyYAML
python-box
langchain
langchain-core
pydantic
weaviate-client
langchain-community
langchain_ollama
langchain_weaviate
langgraph